pyproject.toml
[project]
name = "telegram-nsfw-detector-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "aiogram (>=3.21.0,<4.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "pathlib (>=1.0.1,<2.0.0)",
    "transformers (>=4.55.0,<5.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "opencv-python (>=4.12.0.88,<5.0.0.0)",
    "torch (>=2.8.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
